requirements.txt
Flask==3.0.1
pymongo==4.6.2
python-dotenv==1.0.1
waitress==3.0.0